Implement "setResult" to allow startActivityForResult work in AAOS
Note: current.ignore is required in this CL as we are adding an upper bound to the type of objects returned in CarContext#getCarService().
This is not really a breaking change, because any client using this method properly should continue to compile without changes.

Relnote: Added 'setCarAppResult()' to enable using templated apps 'for result' in AAOS
Bug: 190653108
Test: ./gradlew :car:app:app-automotive:test
Change-Id: I37741edda17f20e598774e8735a4b69cde6c593c