Fixed a memory leak due to the ICarApp binder outliving its parent CarAppService
When the car host unbinds from an app, the ICarApp binder stays around until the
host reference to the binder also gets cleaned up. This often means that the CarAppService
is destroyed first but being held around by the binder being an inner class.

This change makes the binder a static inner class, and we explicitly remove its
reference to the outer CarAppService when the service is destroyed.

Bug: 203594731
Bug: 213378655
Test: Verified with the sample showcase app that a leak is no longer reported
Relnote: Fixed a memory leak in `CarAppService` when the car host unbinds.
Change-Id: I5c9caf05797ea257ef21cbd96dab5fdfbc85a70e