Use ConstraintManager to control the number of items shown in MiscTemplateDemoScreen.
Previously, we would only show 6 items on the MiscTemplateDemo screen regardless of the size of the device. Now we can use the ContraintManager to control how many items will be shown on one page in the MiscTemplateDemoScreen. The idea is that we want to show as many as items on the screen, but we will show the More button if necessary.

Test: Manual Testing with Polestar2 Emulator and Automotive Emulator.
Bug: https://buganizer.corp.google.com/issues/215910755
Relnote: Use the ConstraintManager to control how many items we want to show in one screen.

Screenshot1: https://screenshot.googleplex.com/5Quh8Lf6ptVRKrm
Screenshot2: https://screenshot.googleplex.com/7xNgz6qEioSzvGk

Change-Id: I45fc2e1008fff9067a0d80ba7c021e19b0760e51