Add support for CarIconSpan in the PaneTemplate title
Right now the PaneTemplate doesn't allow CarIconSpans in the title, so
we will see a crash whenever we click a row with icon in the
PlaceListTempateDemo. This CL will add the support for the CarIconSpan
in the CAL, but we need to update the HeaderView in the host side in order to show the
CarIcon in the title.

Test: Manual testing and Unit tests.
Relnote: Add support for CarIconSpans in the PaneTemplate title
Bug: 234631244
Change-Id: Ia1ee0a4ab2e16a4fc29f5714ec29a4c72e3ccab2