Support default action for alerters when timeout is reached
We want to enable the developers to set a default action to be triggered
when the alerter times out. This default action can be different from
the primary action so we should have a new flag for it.

Relnote: Support default action for alerters when timeout is reached
Bug: 237120362
Demo: See the buganizer
Test: Manual testing & Unit tests
Change-Id: I5ad7085f163b8ea5b003a35a97242ac228d4d397