Add android:configChanges="uiMode" to sample apps' manifest to restrict activity recreation during day/night mode change
A day/night mode change causes the CarAppActivity to be destroyed and recreated. This in turn causes the app to unbind and rebind to the host and visually manifests as a black screen flicker. This change prevents the activity from being recreated during uiMode config changes. Additionally, apps can listen to the same configuration changes via Session#onCarConfigurationChanged.

Test: Manual
Bug: 264227000
Change-Id: I6e57f162e71a6a42bd2f195ebc56810bf631db35